Sanitize version ids

This function replaces periods with under scores. The different versions of the data standard are stored in folders with their respective names; however, having periods in folder names can cause problems on certain operating systems and makes it more difficult to parse file extensions.

Usage

sanitize_version(version)

Arguments

version

Character. Version identifier.

Value

Character. Version identifier with no periods.

Examples

sanitize_version("v.1.1.0")
#> [1] "v_1_1_0"